backend/requirements.txt
fastapi
uvicorn[standard]
browser-use
playwright
langgraph
langchain-core
langchain-openai
langchain-google-genai
elevenlabs
openai
google-genai
supermemory
python-dotenv
pydantic
aiohttp
websocket-client
aiortc
opencv-python-headless
pytest
